11.0

+++ extra/xalloc.c
@@ -226,7 +226,7 @@ size_t xdbg_alloc_align(size_t size) {
         g_xdbg_pagesize = getpagesize();
         #endif
         if(!g_xdbg_pagesize) g_xdbg_pagesize = 4096;
-        if(XDBG_ALLOC_VERBOSE) fprintf(stdout, "# g_xdbg_pagesize %u (called by %p)\n", g_xdbg_pagesize, __builtin_extract_return_addr(__builtin_return_address(0)));
+        if(XDBG_ALLOC_VERBOSE) fprintf(stdout, "# g_xdbg_pagesize %u (called by %p)\n", (uint32_t)g_xdbg_pagesize, __builtin_extract_return_addr(__builtin_return_address(0)));
     }
     size = (size + (g_xdbg_pagesize - 1)) & (~(g_xdbg_pagesize - 1));
     return(size);
@@ -574,7 +574,7 @@ void xdbg_realloc_page_guard(void *ptr, size_t new_size) {
         g_xdbg_page_guard, &tmp)) XDBG_ALLOC_ERROR;    // post PAGE_GUARD
     #endif
 
-    if(XDBG_ALLOC_VERBOSE) fprintf(stdout, "# realloc_page_guard(%p, %u, %u) (called by %p)\n", ptr, new_size, total_size, __builtin_extract_return_addr(__builtin_return_address(0)));
+    if(XDBG_ALLOC_VERBOSE) fprintf(stdout, "# realloc_page_guard(%p, %u, %u) (called by %p)\n", ptr, (uint32_t)new_size, (uint32_t)total_size, __builtin_extract_return_addr(__builtin_return_address(0)));
 }
 
 
@@ -589,7 +589,7 @@ void *xdbg_malloc( size_t size ) {
 
     xdbg_alloc_init(&original, &size, &xsize);
 
-    if(XDBG_ALLOC_VERBOSE) fprintf(stdout, "# malloc(%u) -> %u %u (called by %p)\n", original, size, xsize, __builtin_extract_return_addr(__builtin_return_address(0)));
+    if(XDBG_ALLOC_VERBOSE) fprintf(stdout, "# malloc(%u) -> %u %u (called by %p)\n", (uint32_t)original, (uint32_t)size, (uint32_t)xsize, __builtin_extract_return_addr(__builtin_return_address(0)));
 
     if(XDBG_HEAPVALIDATE) {
         if(!HeapValidate(GetProcessHeap(), 0, NULL)) XDBG_ALLOC_ERROR;
@@ -640,7 +640,7 @@ void *xdbg_malloc( size_t size ) {
 
     if(XDBG_ALLOC_VERBOSE)
         fprintf(stdout, "# %p (%p -> %p) = malloc(%u / %u / %u) (called by %p)\n",
-            mem, XDBG_ALLOC_MEM2PTR(mem), (unsigned char *)mem + xsize, original, size, xsize, __builtin_extract_return_addr(__builtin_return_address(0)));
+            mem, XDBG_ALLOC_MEM2PTR(mem), (unsigned char *)mem + xsize, (uint32_t)original, (uint32_t)size, (uint32_t)xsize, __builtin_extract_return_addr(__builtin_return_address(0)));
     return(XDBG_ALLOC_MEM2PTR(mem));
 }
 
@@ -663,7 +663,7 @@ void *xdbg_realloc( void *ptr, size_t size ) {
         return(real_realloc(ptr, size));
     }
 
-    if(XDBG_ALLOC_VERBOSE) fprintf(stdout, "# realloc(%p, %u) (called by %p)\n", ptr, size, __builtin_extract_return_addr(__builtin_return_address(0)));
+    if(XDBG_ALLOC_VERBOSE) fprintf(stdout, "# realloc(%p, %u) (called by %p)\n", ptr, (uint32_t)size, __builtin_extract_return_addr(__builtin_return_address(0)));
 
     if(!ptr) return(xdbg_malloc(size));
 
@@ -679,7 +679,7 @@ void *xdbg_realloc( void *ptr, size_t size ) {
 
     xdbg_alloc = XDBG_ALLOC_PTR2INFO(ptr);
     if(xdbg_alloc->size >= size) {  // aligned_allocated versus aligned_requested
-        if(XDBG_ALLOC_VERBOSE) fprintf(stdout, "# realloc reused same buffer %p (%u %u) (called by %p)\n", ptr, size, xdbg_alloc->size, __builtin_extract_return_addr(__builtin_return_address(0)));
+        if(XDBG_ALLOC_VERBOSE) fprintf(stdout, "# realloc reused same buffer %p (%u %u) (called by %p)\n", ptr, (uint32_t)size, (uint32_t)xdbg_alloc->size, __builtin_extract_return_addr(__builtin_return_address(0)));
         xdbg_alloc->original = original;
         //memset(ptr, 0, xdbg_alloc->size - original);  // takes really lot of time!
         //xdbg_realloc_page_guard(ptr, size);
@@ -714,7 +714,7 @@ void *xdbg_realloc( void *ptr, size_t size ) {
         void    *old_mem = XDBG_ALLOC_PTR2MEM(ptr); // not a problem for realloc_to_file
         size_t  tmp = XDBG_ALLOC_MEM2INFO(new_mem)->xsize;
         fprintf(stdout, "# %p (%p -> %p) = realloc(%p, %u / %u / %u) (called by %p)\n",
-            new_mem, new_ptr, (unsigned char *)new_mem + tmp, old_mem, original, size, tmp, __builtin_extract_return_addr(__builtin_return_address(0)));
+            new_mem, new_ptr, (unsigned char *)new_mem + tmp, old_mem, (uint32_t)original, (uint32_t)size, (uint32_t)tmp, __builtin_extract_return_addr(__builtin_return_address(0)));
     }
     if(realloc_to_file) {
         fseek(fd, 0, SEEK_SET);
@@ -731,7 +731,7 @@ void *xdbg_realloc( void *ptr, size_t size ) {
 void *xdbg_calloc( size_t num, size_t size ) {
     if(!XDBG_ALLOC_ACTIVE) return(real_calloc(num, size));
 
-    if(XDBG_ALLOC_VERBOSE) fprintf(stdout, "# calloc(%u, %u) (called by %p)\n", num, size, __builtin_extract_return_addr(__builtin_return_address(0)));
+    if(XDBG_ALLOC_VERBOSE) fprintf(stdout, "# calloc(%u, %u) (called by %p)\n", (uint32_t)num, (uint32_t)size, __builtin_extract_return_addr(__builtin_return_address(0)));
 
     uint64_t    tmp;
     tmp = (uint64_t)num * (uint64_t)size;
